All images download using SDWebImage, providing caching and better stability. Code cleanup. Removed episode progress meter. Layout adjustments.

+++ SickBeard/Classes/SBShow.m
@@ -82,7 +82,7 @@
 - (NSString*)sanitizedShowName {
 	if ([self.showName rangeOfCharacterFromSet:[NSCharacterSet characterSetWithCharactersInString:@"()"]].location != NSNotFound) {
 		NSString *foundData = @"";
-		int left, right;
+		NSUInteger left, right;
 
 		NSScanner *scanner = [NSScanner scannerWithString:self.showName];
 		[scanner scanUpToString:@"(" intoString:nil];
